requirements.txt
Django==4.1.7
djangorestframework==3.14.0
drf-yasg==1.21.5
snapshottest==0.6.0
factory-boy==3.2.1
